requirements.txt
kokoro
kokoro-onnx
pip-chill
pyaudio
pymultidictionary
serial
sounddevice
spacy-curated-transformers
sudachidict-core
tf-keras
tts
vosk
